+++ lib/Event/Model/AssetDeleteInfoEvent.php
@@ -13,17 +13,15 @@
  *  @license    http://www.pimcore.org/license     GPLv3 and PCL
  */
 
-namespace Pimcore\Bundle\AdminBundle\Event\Model;
-
-use Pimcore\Bundle\AdminBundle\Event\Traits\ElementDeleteInfoEventTrait;
-use Pimcore\Event\Model\DocumentEvent;
+namespace Pimcore\Event\Model;
 
 /**
- * Class DocumentDeleteInfoEvent
+ * Class AssetDeleteInfoEvent
  *
  * @package Pimcore\Event\Model
+ *
+ * @deprecated and will be removed in Pimcore 11. Use Pimcore\Bundle\AdminBundle\Event\Model\AssetDeleteInfoEvent instead
  */
-class DocumentDeleteInfoEvent extends DocumentEvent implements ElementDeleteInfoEventInterface
+class AssetDeleteInfoEvent extends \Pimcore\Bundle\AdminBundle\Event\Model\AssetDeleteInfoEvent
 {
-    use ElementDeleteInfoEventTrait;
 }
